Use hostname as server name instead of MAC address
The DGRAM key exchange broke naming the server based on the MAC of the
socket, since the local socket is unbound (inaddr_any).

This isn't worth fixing: it makes a lot more sense for the client to use
the MAC as its name than it does for the server. Just use the hostname.