Add support for streaming over raw ethernet frames
A neat benefit of the UDP port multiplexing is that almost the same code
can support raw ethernet frames. The key exchange works too.

The wart is on the client side: AF_PACKET sockets don't support connect(),
so write() has to become sendto() with an explicit sockaddr in the dgram
tx code. It might be less ugly to just never connect()... but for now it
still does in the UDP case.

I have no idea why anybody would want to do this... but now you can!