Use liburcu
Replace the kludgey probing hashtable with rculfhash. RCU manages the
lifetime of the client objects, all lookups go thorugh the lfht and no
refs are held across the event loops so we can quiesce frequently.

In theory we could store the pointer to the client struct in the
epoll_event data and avoid the hashtable lookup for the STREAM case,
but that complicates tearing down the client and I'm skeptical it
would make a big difference.

The cookies now begin at a random value.